Refactor NextJob to client.NextJob
