+++ job-service/job-service_test.go
@@ -7,18 +7,17 @@ import (
 	"log"
 	"net/http"
 	"net/http/httptest"
-	"net/url"
 	"os"
-	"sync"
 	"testing"
 	"time"
 
 	"bou.ke/monkey"
 	"github.com/go-test/deep"
-	"github.com/m-lab/etl-gardener/job-service"
-	jobservice "github.com/m-lab/etl-gardener/job-service"
-	"github.com/m-lab/etl-gardener/tracker"
+
 	"github.com/m-lab/go/rtx"
+
+	"github.com/m-lab/etl-gardener/job-service"
+	"github.com/m-lab/etl-gardener/tracker"
 )
 
 func init() {
@@ -35,7 +34,7 @@ func TestService_NextJob(t *testing.T) {
 	defer monkey.Unpatch(time.Now)
 
 	start := time.Date(2011, 2, 3, 5, 6, 7, 8, time.UTC)
-	svc, _ := jobservice.NewJobService(nil, "fake-bucket", start)
+	svc, _ := job.NewJobService(nil, "fake-bucket", start)
 	j := svc.NextJob()
 	w, err := tracker.Job{Bucket: "fake-bucket", Experiment: "ndt", Datatype: "ndt5", Date: start.Truncate(24 * time.Hour)}.Target("gs://fakebucket/ndt/ndt5")
 	rtx.Must(err, "")
@@ -161,79 +160,5 @@ func TestEarlyWrapping(t *testing.T) {
 				t.Error(err)
 			}
 		}
-	}
-}
-
-type fakeGardener struct {
-	t *testing.T // for logging
-
-	lock       sync.Mutex
-	jobs       []tracker.JobWithTarget
-	heartbeats int
-	updates    int
-}
-
-func (g *fakeGardener) AddJob(job tracker.Job, target string) error {
-	jt, err := job.Target(target)
-	if err != nil {
-		return err
-	}
-	g.jobs = append(g.jobs, jt)
-	return nil
-}
-
-func (g *fakeGardener) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	rtx.Must(r.ParseForm(), "bad request")
-	if r.Method != http.MethodPost {
-		log.Fatal("Should be POST") // Not t.Fatal because this is asynchronous.
-	}
-	g.lock.Lock()
-	g.lock.Unlock()
-	switch r.URL.Path {
-	case "/job":
-		if len(g.jobs) < 1 {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		j := g.jobs[0]
-		g.jobs = g.jobs[1:]
-		w.Write(j.Marshal())
-	case "/heartbeat":
-		g.t.Log(r.URL.Path, r.URL.Query())
-		g.heartbeats++
-
-	case "/update":
-		g.t.Log(r.URL.Path, r.URL.Query())
-		g.updates++
-
-	default:
-		log.Fatal(r.URL) // Not t.Fatal because this is asynchronous.
-	}
-}
-
-func TestJobClient(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	// set up a fake gardener service.
-	fg := fakeGardener{t: t, jobs: make([]tracker.JobWithTarget, 0)}
-	spec := tracker.NewJob(
-		"foobar", "ndt", "ndt5", time.Date(2019, 01, 01, 0, 0, 0, 0, time.UTC))
-	rtx.Must(fg.AddJob(spec, "a.b.c"), "add job")
-	gardener := httptest.NewServer(&fg)
-	defer gardener.Close()
-	gURL, err := url.Parse(gardener.URL)
-	rtx.Must(err, "bad url")
-
-	j, err := jobservice.NextJob(ctx, *gURL)
-	rtx.Must(err, "next job")
-
-	if j.Path() != "gs://foobar/ndt/ndt5/2019/01/01/" {
-		t.Error(j.Path())
-	}
-
-	j, err = jobservice.NextJob(ctx, *gURL)
-	if err.Error() != "Internal Server Error" {
-		t.Fatal("Should be internal server error", err)
 	}
 }
